+++ src/fauxmoESP.cpp
@@ -35,16 +35,21 @@ void fauxmoESP::_sendUDPResponse(unsigned int device_id) {
     fauxmoesp_device_t device = _devices[device_id];
     DEBUG_MSG_FAUXMO("[FAUXMO] UDP response for device #%d (%s)\n", _current, device.name);
 
+    char buffer[16];
+    IPAddress ip = WiFi.localIP();
+    sprintf(buffer, "%d.%d.%d.%d", ip[0], ip[1], ip[2], ip[3]);
+
     char response[strlen(UDP_TEMPLATE) + 40];
     sprintf_P(response, UDP_TEMPLATE,
-        WiFi.localIP().toString().c_str(),
-        _base_port + _current, device.uuid, device.uuid
+        buffer,
+        _base_port + _current,
+        device.uuid,
+        device.uuid
     );
 
-    WiFiUDP udpClient;
-    udpClient.beginPacket(_remoteIP, _remotePort);
-    udpClient.write(response);
-    udpClient.endPacket();
+    _udp.beginPacket(_remoteIP, _remotePort);
+    _udp.write(response);
+    _udp.endPacket();
 
 }
 
@@ -71,15 +76,16 @@ void fauxmoESP::_handleUDPPacket(IPAddress remoteIP, unsigned int remotePort, ui
 
     if (!_enabled) return;
 
-    //DEBUG_MSG_FAUXMO("[FAUXMO] Got UDP packet from %s\n", remoteIP.toString().c_str());
-
     data[len] = 0;
-    String content = String((char *) data);
 
-    if (content.indexOf(UDP_SEARCH_PATTERN) == 0) {
-        if (content.indexOf(UDP_DEVICE_PATTERN) > 0) {
+    if (strstr((char *) data, UDP_SEARCH_PATTERN) == (char *) data) {
+        if (strstr((char *) data, UDP_DEVICE_PATTERN) != NULL) {
 
-            DEBUG_MSG_FAUXMO("[FAUXMO] Search request from %s\n", remoteIP.toString().c_str());
+            #ifdef DEBUG_FAUXMO
+                char buffer[16];
+                sprintf(buffer, "%d.%d.%d.%d", remoteIP[0], remoteIP[1], remoteIP[2], remoteIP[3]);
+                DEBUG_MSG_FAUXMO("[FAUXMO] Search request from %s\n", buffer);
+            #endif
 
             // Set hits to false
             for (unsigned int i = 0; i < _devices.size(); i++) {
@@ -97,94 +103,34 @@ void fauxmoESP::_handleUDPPacket(IPAddress remoteIP, unsigned int remotePort, ui
 
 }
 
-void fauxmoESP::_sendTCPPacket(AsyncClient *client, const char * response) {
-    char buffer[strlen(HEADERS) + strlen(response) + 10];
-    sprintf_P(buffer, HEADERS, strlen(response), response);
-    client->write(buffer);
-}
+void fauxmoESP::_handleSetup(AsyncWebServerRequest *request, unsigned int device_id) {
 
-void fauxmoESP::_handleTCPPacket(unsigned int device_id, AsyncClient *client, void *data, size_t len) {
+    if (!_enabled) return;
 
-    ((char * )data)[len] = 0;
-    String content = String((char *) data);
+    DEBUG_MSG_FAUXMO("[FAUXMO] Device #%d /setup.xml\n", device_id);
+    _devices[device_id].hit = true;
 
     fauxmoesp_device_t device = _devices[device_id];
-
-    if (content.indexOf("GET /setup.xml") == 0) {
-
-        DEBUG_MSG_FAUXMO("[FAUXMO] /setup.xml response for device #%d (%s)\n", device_id, device.name);
-
-        _devices[device_id].hit = true;
-        char response[strlen(SETUP_TEMPLATE) + 50];
-        sprintf_P(response, SETUP_TEMPLATE, device.name, device.uuid);
-        _sendTCPPacket(client, response);
-        client->close();
-
-    }
-
-    if (content.indexOf("POST /upnp/control/basicevent1") == 0) {
-
-        if (content.indexOf("<BinaryState>0</BinaryState>") > 0) {
-            if (_callback) _callback(device_id, device.name, false);
-        }
-
-        if (content.indexOf("<BinaryState>1</BinaryState>") > 0) {
-            if (_callback) _callback(device_id, device.name, true);
-        }
-
-        _sendTCPPacket(client, "");
-        client->close();
-
-    }
+    char response[strlen(SETUP_TEMPLATE) + 50];
+    sprintf_P(response, SETUP_TEMPLATE, device.name, device.uuid);
+    request->send(200, "text/xml", response);
 
 }
 
-AcConnectHandler fauxmoESP::_getTCPClientHandler(unsigned int device_id) {
+void fauxmoESP::_handleContent(AsyncWebServerRequest *request, unsigned int device_id, char * content) {
 
-    return [this, device_id](void *s, AsyncClient * client) {
+    if (!_enabled) return;
 
-        if (!_enabled) return;
+    DEBUG_MSG_FAUXMO("[FAUXMO] Device #%d /upnp/control/basicevent1\n", device_id);
+    fauxmoesp_device_t device = _devices[device_id];
 
-        for (int i = 0; i < TCP_MAX_CLIENTS; i++) {
-            if (!_clients[i] || !_clients[i]->connected()) {
+    if (strstr(content, "<BinaryState>0</BinaryState>") != NULL) {
+        if (_callback) _callback(device_id, device.name, false);
+    }
 
-                _clients[i] = client;
-
-                client->onAck([this, i](void *s, AsyncClient *c, size_t len, uint32_t time) {
-                    //DEBUG_MSG_FAUXMO("[FAUXMO] Got ack for client %i len=%u time=%u\n", i, len, time);
-                }, 0);
-
-                client->onData([this, i, device_id](void *s, AsyncClient *c, void *data, size_t len) {
-                    //DEBUG_MSG_FAUXMO("[FAUXMO] Got data from client %i len=%i\n", i, len);
-                    _handleTCPPacket(device_id, c, data, len);
-                }, 0);
-
-                client->onDisconnect([this, i](void *s, AsyncClient *c) {
-                    //DEBUG_MSG_FAUXMO("[FAUXMO] Disconnect for client %i\n", i);
-                    _clients[i]->free();
-                }, 0);
-                client->onError([this, i](void *s, AsyncClient *c, int8_t error) {
-                    //DEBUG_MSG_FAUXMO("[FAUXMO] Error %s (%i) on client %i\n", c->errorToString(error), error, i);
-                }, 0);
-                client->onTimeout([this, i](void *s, AsyncClient *c, uint32_t time) {
-                    //DEBUG_MSG_FAUXMO("[FAUXMO] Timeout on client %i at %i\n", i, time);
-                    c->close();
-                }, 0);
-
-                return;
-
-            }
-        }
-
-        DEBUG_MSG_FAUXMO("[FAUXMO] Rejecting client - Too many connections already.\n");
-
-        // We cannot accept this connection at the moment
-        client->onDisconnect([](void *s, AsyncClient *c) {
-            delete(c);
-        });
-        client->stop();
-
-    };
+    if (strstr(content, "<BinaryState>1</BinaryState>") != NULL) {
+        if (_callback) _callback(device_id, device.name, true);
+    }
 
 }
 
@@ -202,12 +148,27 @@ void fauxmoESP::addDevice(const char * device_name) {
     new_device.uuid = strdup(uuid);
 
     // TCP Server
-    new_device.server = new AsyncServer(_base_port + device_id);
-    new_device.server->onClient(_getTCPClientHandler(device_id), 0);
+    new_device.server = new AsyncWebServer(_base_port + device_id);
+    new_device.server->on("/setup.xml", HTTP_GET, [this, device_id](AsyncWebServerRequest *request){
+        _handleSetup(request, device_id);
+    });
+    new_device.server->on("/upnp/control/basicevent1", HTTP_POST,
+        [](AsyncWebServerRequest *request) {
+            request->send(200);
+        },
+        NULL,
+        [this, device_id](AsyncWebServerRequest *request, uint8_t *data, size_t len, size_t index, size_t total) {
+            data[len] = 0;
+            _handleContent(request, device_id, (char *) data);
+        }
+    );
+
     new_device.server->begin();
 
     // Attach
     _devices.push_back(new_device);
+
+    DEBUG_MSG_FAUXMO("[FAUXMO] Device '%s' added (#%d)\n", device_name, device_id);
 
 }
 
@@ -229,6 +190,11 @@ void fauxmoESP::handle() {
         }
     }
 
+}
+
+void fauxmoESP::enable(bool enable) {
+    DEBUG_MSG_FAUXMO("[FAUXMO] %s\n", enable ? "Enabled" : "Disabled");
+    _enabled = enable;
 }
 
 fauxmoESP::fauxmoESP(unsigned int port) {